+++ elements/user-picker/examples/20-multi-with-external-users.tsx
@@ -16,7 +16,6 @@ export default class Example extends React.Component<{}> {
             onInputChange={onInputChange}
             noOptionsMessage={() => null}
             isMulti
-            allowEmail
           />
         )}
       </ExampleWrapper>