+++ elements/user-picker/src/__tests__/unit/components/AddOptionAvatarSpec.tsx
@@ -1,4 +1,5 @@
 import InviteTeamIcon from '@atlaskit/icon/glyph/invite-team';
+import { N200 } from '@atlaskit/theme/colors';
 import { shallow } from 'enzyme';
 import React from 'react';
 import {
@@ -22,6 +23,22 @@ describe('AddOptionAvatar', () => {
       label: 'Invite',
       size: 'small',
       primaryColor: 'white',
+    });
+  });
+
+  it('should render add avatar Icon when it is a suggestion', () => {
+    const component = shallowAddOptionAvatar({
+      label: 'Invite',
+      size: 'small',
+      suggestion: true,
+    });
+
+    const inviteIcon = component.find(InviteTeamIcon);
+    expect(inviteIcon).toHaveLength(1);
+    expect(inviteIcon.props()).toMatchObject({
+      label: 'Invite',
+      size: 'small',
+      primaryColor: N200,
     });
   });
 });