+++ elements/user-picker/src/__tests__/unit/components/EmailOptionSpec.tsx
@@ -18,6 +18,14 @@ describe('EmailOption', () => {
     lozenge: 'EMAIL',
   };
 
+  const suggestedEmail: Email = {
+    type: EmailType,
+    id: 'test@test.com',
+    name: 'test@test.com',
+    lozenge: 'EMAIL',
+    suggestion: true,
+  };
+
   it('should render AvatarItemOption with default message', () => {
     const component = shallowEmailOption({
       isSelected: false,
@@ -31,6 +39,29 @@ describe('EmailOption', () => {
     const avatarItemOption = message.find(AvatarItemOption);
     expect(avatarItemOption.props()).toMatchObject({
       avatar: <AddOptionAvatar label="Invite" />,
+      primaryText: 'test@test.com',
+      secondaryText: 'Invite',
+      lozenge: {
+        text: 'EMAIL',
+      },
+    });
+  });
+
+  it('should render AvatarItemOption for suggested emails', () => {
+    const component = shallowEmailOption({
+      isSelected: false,
+      email: suggestedEmail,
+      emailValidity: 'VALID',
+    });
+
+    const formattedMessage = component.find(FormattedMessage);
+    expect(formattedMessage).toHaveLength(1);
+    const message = renderProp(formattedMessage, 'children', 'Invite');
+    const avatarItemOption = message.find(AvatarItemOption);
+    expect(avatarItemOption.props()).toMatchObject({
+      avatar: (
+        <AddOptionAvatar label="Invite" suggestion={true} size={'medium'} />
+      ),
       primaryText: 'test@test.com',
       secondaryText: 'Invite',
       lozenge: {