+++ elements/user-picker/src/components/AddOptionAvatar.tsx
@@ -1,9 +1,10 @@
 import InviteTeamIcon from '@atlaskit/icon/glyph/invite-team';
-import { N50 } from '@atlaskit/theme/colors';
+import MentionIcon from '@atlaskit/icon/glyph/mention';
+import { N50, N200 } from '@atlaskit/theme/colors';
 import React from 'react';
 import styled from 'styled-components';
 
-const AddOptionAvatarWrapper = styled.span`
+const AddOptionAvatarWrapper = styled.span<{ suggestion?: boolean }>`
   color: black;
   padding: 2px;
 
@@ -11,21 +12,34 @@ const AddOptionAvatarWrapper = styled.span`
     background-color: ${N50};
     border-radius: 50%;
   }
+
+  ${({ suggestion }) => suggestion && `border-radius: 50%;`}
 `;
 
 export type AddOptionAvatarProps = {
   label: string;
-  size?: 'small' | 'large';
+  size?: 'small' | 'medium' | 'large';
+  suggestion?: boolean;
+  invalidOption?: boolean;
 };
 
-export const AddOptionAvatar: React.StatelessComponent<AddOptionAvatarProps> = ({
+export const AddOptionAvatar: React.FunctionComponent<AddOptionAvatarProps> = ({
   size,
   label,
-}) => (
-  <AddOptionAvatarWrapper>
-    <InviteTeamIcon label={label} size={size} primaryColor="white" />
-  </AddOptionAvatarWrapper>
-);
+  suggestion,
+  invalidOption,
+}) => {
+  const Icon = invalidOption ? MentionIcon : InviteTeamIcon;
+  return (
+    <AddOptionAvatarWrapper suggestion={suggestion}>
+      <Icon
+        label={label}
+        size={size}
+        primaryColor={suggestion ? N200 : 'white'}
+      />
+    </AddOptionAvatarWrapper>
+  );
+};
 
 AddOptionAvatar.defaultProps = {
   size: 'large',