+++ analytics/analytics-namespaced-context/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/analytics-namespaced-context
 
+## 5.1.4
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 5.1.3
 
 ### Patch Changes