+++ confluence/tree/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/tree
 
+## 8.1.3
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 8.1.2
 
 ### Patch Changes