+++ css-packs/css-reset/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/css-reset
 
+## 6.0.5
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 6.0.4
 
 ### Patch Changes