+++ design-system/avatar-group/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/avatar-group
 
+## 8.0.12
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 8.0.11
 
 ### Patch Changes