+++ design-system/breadcrumbs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/breadcrumbs
 
+## 10.0.10
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 10.0.9
 
 ### Patch Changes