+++ design-system/checkbox/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/checkbox
 
+## 11.0.8
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 11.0.7
 
 ### Patch Changes