+++ design-system/drawer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/drawer
 
+## 6.0.7
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 6.0.6
 
 ### Patch Changes