+++ design-system/empty-state/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/empty-state
 
+## 7.1.5
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 7.1.4
 
 ### Patch Changes