+++ design-system/feedback-collector/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/feedback-collector
 
+## 6.1.5
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 6.1.4
 
 ### Patch Changes