+++ design-system/flag/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/flag
 
+## 14.0.8
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 14.0.7
 
 ### Patch Changes