+++ design-system/layer-manager/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/layer-manager
 
+## 9.0.4
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 9.0.3
 
 ### Patch Changes