+++ design-system/modal-dialog/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/modal-dialog
 
+## 11.2.4
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 11.2.3
 
 ### Patch Changes