+++ design-system/motion/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/motion
 
+## 0.4.6
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 0.4.5
 
 ### Patch Changes