+++ design-system/multi-select/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/multi-select
 
+## 15.0.5
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 15.0.4
 
 ### Patch Changes