+++ design-system/navigation-next/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/navigation-next
 
+## 9.0.9
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 9.0.8
 
 ### Patch Changes