+++ design-system/page-layout/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/page-layout
 
+## 0.10.10
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 0.10.9
 
 ### Patch Changes