+++ design-system/progress-tracker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/progress-tracker
 
+## 7.0.6
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 7.0.5
 
 ### Patch Changes