+++ design-system/single-select/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/single-select
 
+## 10.0.4
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 10.0.3
 
 ### Patch Changes