+++ design-system/size-detector/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/size-detector
 
+## 10.0.3
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 10.0.2
 
 ### Patch Changes