+++ design-system/spinner/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/spinner
 
+## 15.0.6
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 15.0.5
 
 ### Patch Changes