+++ design-system/table-tree/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/table-tree
 
+## 9.0.7
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 9.0.6
 
 ### Patch Changes