+++ design-system/tag/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/tag
 
+## 11.0.5
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 11.0.4
 
 ### Patch Changes