+++ design-system/toggle/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/toggle
 
+## 12.0.1
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 12.0.0
 
 ### Major Changes