+++ design-system/tooltip/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/tooltip
 
+## 17.1.2
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 17.1.1
 
 ### Patch Changes