+++ editor/adf-schema/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/adf-schema
 
+## 13.3.1
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 13.3.0
 
 ### Minor Changes