+++ editor/editor-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/editor-core
 
+## 134.0.1
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+- Updated dependencies
+
 ## 134.0.0
 
 ### Major Changes