+++ editor/editor-markdown-transformer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-markdown-transformer
 
+## 4.1.4
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 4.1.3
 
 ### Patch Changes