+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @atlaskit/editor-mobile-bridge
 
+## 20.1.2
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+- Updated dependencies
+
+## 20.1.1
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 20.1.0
 
 ### Minor Changes