+++ editor/editor-shared-styles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-shared-styles
 
+## 1.1.7
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 1.1.6
 
 ### Patch Changes