+++ editor/editor-wikimarkup-transformer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-wikimarkup-transformer
 
+## 8.3.5
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 8.3.4
 
 ### Patch Changes