+++ editor/embedded-document/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/embedded-document
 
+## 0.7.11
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 0.7.10
 
 ### Patch Changes