+++ elements/date/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/date
 
+## 0.8.4
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 0.8.3
 
 ### Patch Changes