+++ elements/mention/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/mention
 
+## 19.1.1
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 19.1.0
 
 ### Minor Changes