+++ elements/share/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/share
 
+## 0.24.3
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 0.24.2
 
 ### Patch Changes