+++ elements/status/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/status
 
+## 0.10.6
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 0.10.5
 
 ### Patch Changes