+++ elements/task-decision/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/task-decision
 
+## 17.0.9
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 17.0.8
 
 ### Patch Changes