+++ help/right-side-panel/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/right-side-panel
 
+## 0.4.4
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 0.4.3
 
 ### Patch Changes