+++ jira/color-picker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/color-picker
 
+## 1.1.1
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 1.1.0
 
 ### Minor Changes