+++ media/media-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/media-client
 
+## 12.1.1
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+- Updated dependencies
+
 ## 12.1.0
 
 ### Minor Changes