+++ media/media-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-common
 
+## 2.4.1
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 2.4.0
 
 ### Minor Changes