+++ media/media-filmstrip/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/media-filmstrip
 
+## 41.0.1
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+- Updated dependencies
+
 ## 41.0.0
 
 ### Patch Changes