+++ media/media-table/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/media-table
 
+## 8.0.1
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+- Updated dependencies
+
 ## 8.0.0
 
 ### Patch Changes