+++ media/media-test-helpers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-test-helpers
 
+## 28.4.1
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 28.4.0
 
 ### Minor Changes