+++ navigation/atlassian-navigation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/atlassian-navigation
 
+## 0.12.4
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 0.12.3
 
 ### Patch Changes