+++ navigation/atlassian-notifications/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/atlassian-notifications
 
+## 0.2.4
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 0.2.3
 
 ### Patch Changes