+++ teamwork-ecosystem/outbound-auth-flow-client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/outbound-auth-flow-client
 
+## 3.2.1
+
+### Patch Changes
+
+- [`d3265f19be`](https://bitbucket.org/atlassian/atlassian-frontend/commits/d3265f19be) - Transpile packages using babel rather than tsc
+
 ## 3.2.0
 
 ### Minor Changes